database: Move the writer into a separate file.
